// components/proveedores/types.ts
export interface Provider {
    id: string;
    nombreProveedor: string;
    contactoPrincipal: string;
    telefono: string;
    correo: string;
    direccion: string;
    idFiscal: string;
    condicionesPago: string;
    plazoEntrega: string;
    productosSuministra: string;
    historialPedidos: string;
    calificacion: number;
    notas: string;
    estado: string;
    fechaRegistro: string;
    docAdjunta: string;
    categoriaProveedor: string;
    contactoSecundario: string;
  }
  